Fix Bug that prevents marking events as published
