+++ composer.json
@@ -41,6 +41,10 @@
 	  "ext-mbstring" : "*",
       "ext-intl": "*"
     },
+	"require-dev" : {
+		"squizlabs/php_codesniffer": "^3.7.2",
+		"phpunit/phpunit": "9.6.10"
+	},
 	"autoload" : {
 		"psr-4" : {
 			"ruhrpottmetaller\\" : "ruhrpottmetaller"
